app/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
djangorestframework
channels
channels_redis
uvicorn[standard]
requests
websockets
wsproto
# UI
django-theme-material-kit==1.0.25
# Deployment
whitenoise==6.6.0
djangorestframework-simplejwt
django-cuser
celery
redis
unidecode
google-api-python-client
google-auth-httplib2
google-auth-oauthlib